pyproject.toml
[project]
name = "twobridgesscenario"
version = "0.1.0"
description = ""
authors = [
    {name = "Matteo Magnini",email = "matteo.magnini@unibo.it"}
]
license = {text = "Apache-2.0"}
requires-python = ">=3.14"
dependencies = [
    "numpy (>=2.4.4,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
